failed result button press case
Added a case to handle button press during display of a failed result. FAILED_RESULT_TIME is simply to ensure the message is displayed that the result has been failed. In terms of functionality nothing changed, as I set this to be the same as CAR_ON_TIME, but it seemed best not to deal with a failed result case using CAR_ON_TIME when the car was never on.